# Conflicts:
#	client/pom.xml
#	core/pom.xml
#	core/src/main/java/com/orientechnologies/orient/core/OConstants.java
#	distributed/pom.xml
#	distribution/pom.xml
#	etl/pom.xml
#	graphdb/pom.xml
#	jdbc/pom.xml
#	jmh-tests/pom.xml
#	lucene/pom.xml
#	object/pom.xml
#	pom.xml
#	server/pom.xml
#	test-commons/pom.xml
#	tests/pom.xml
#	tools/pom.xml

+++ server/src/test/java/com/orientechnologies/orient/test/server/network/http/HttpDocumentTest.java
@@ -144,7 +144,7 @@ public class HttpDocumentTest extends BaseHttpDatabaseTest {
 
     Assert.assertEquals(created.field("name"), "Jay");
     Assert.assertEquals(created.field("surname"), "Miner");
-    Assert.assertEquals(created.field("age"), 0);
+    Assert.assertEquals(created.field("age"), (Integer)0);
     Assert.assertEquals(created.getVersion(), 1);
 
     patch("document/" + getDatabaseName() + "/" + created.getIdentity().toString().substring(1)).payload("{age:1,@version:1}", CONTENT.JSON)
@@ -154,7 +154,7 @@ public class HttpDocumentTest extends BaseHttpDatabaseTest {
 
     Assert.assertEquals(updated.field("name"), "Jay");
     Assert.assertEquals(updated.field("surname"), "Miner");
-    Assert.assertEquals(updated.field("age"), 1);
+    Assert.assertEquals(updated.field("age"), (Integer)1);
     Assert.assertEquals(updated.getVersion(), 2);
   }
